refactor: Alterado nome de package de exceptions
