+++ src/main/java/com/bsg/trustedone/exception/UserAlreadyRegisteredException.java
@@ -1,4 +1,4 @@
-package com.bsg.trustedone.exceptions;
+package com.bsg.trustedone.exception;
 
 public class UserAlreadyRegisteredException extends RuntimeException {
