+++ src/main/java/com/bsg/trustedone/exception/UserLoginException.java
@@ -1,4 +1,4 @@
-package com.bsg.trustedone.exceptions;
+package com.bsg.trustedone.exception;
 
 import lombok.Getter;
